Use external API for default select
