--- packages/liveblocks-react-tiptap/src/toolbar/FloatingToolbarContext.ts
@@ -1,11 +1,0 @@
-import { createContext } from "react";
-
-// This file is separate to avoid circular dependencies
-
-type FloatingToolbarContext = {
-  close: () => void;
-  registerExternal: (id: string) => () => void;
-};
-
-export const FloatingToolbarContext =
-  createContext<FloatingToolbarContext | null>(null);
